Add initial bottom-panel support, not configurable yet
After 5.75 hours of fucking around with GTK, we can safely determine it is
completely devoid of sanity. Not only are positions and sizes entirely
reliant on a signal system that would be put to shame by a bus conductor,
the API to do anything in something resembling a synchronous manner, or
to even "resize now", is entirely private API.

We also learn that GtkPopover should be shot with extreme prejudice, again
using internal API not accessible to us poor peasants, relying on obscene
"support" within GtkWindow. To add further insult to injury we find that
using a shaped window will not work when using a GtkPopover, because we
need to ensure it can merge input child shapes. On the window we can't
get a fix on.

To summarize, I could either have been starting to watch Frodo begin his
treacherous journey to Orthanc by now, having already finished his initial
perilous journey through Middle Earth around the time I realised I was
entirely doomed to deal with this API - or more sensible I could have
taken the time to combat my internal disgust for C++, rationalised my
mentality on the subject and became extremely proficient in the use of
Qt.