Made MPRIS ClientWidgets collapsible
With many players open the media pane can get rather
cluttered. Hence my implementation of a collapse function.

I arranged the collapse/expand button to the very right. Not all
players have close buttons, leading to an inconsistent "column layout",
if the close button is right of the collapse button.